Student mark type changed
Mark propery is supposed to represent average mark, so it can be not integer.

+++ lab1/Student.h
@@ -7,7 +7,7 @@
 
 @interface Student : Human<AvgMarkable>
 
-@property (nonatomic) int mark;
+@property (nonatomic) float mark;
 @property (nonatomic, weak) Group *group;
 
 @end
